requirements.txt
allure-pytest-bdd==2.14.3
allure-python-commons==2.14.3
browserstack-local==1.2.12
browserstack-sdk==1.24.1
PyHamcrest==2.1.0
pytest-bdd==8.1.0
pytest-html==4.1.1
pytest-rerunfailures==15.1
pytest-retry==1.7.0
pytest-variables==3.1.0
pytest==8.4.1
python-dateutil==2.8.2
pytz==2025.2
requests==2.32.4
selenium==4.34.2